+++ python/metatensor_learn/metatensor_learn/nn/__init__.py
@@ -1,7 +1,7 @@
 from ._equivariant import EquivariantTransformation  # noqa: F401
 from ._layer_norm import InvariantLayerNorm, LayerNorm  # noqa: F401
 from ._linear import EquivariantLinear, Linear  # noqa: F401
-from ._module import Module  # noqa: F401
+from ._module import Buffer, Module  # noqa: F401
 from ._module_map import ModuleMap  # noqa: F401
 from ._relu import InvariantReLU, ReLU  # noqa: F401
 from ._sequential import Sequential  # noqa: F401